move: 파일 이동
  - 엔티티를 infrastructure/persistence/entity로 이동
  - enum, 도메인 인터페이스 domain/model로 이동

(#27)
