+++ backend/product-service/src/main/java/com/bidket/product/domain/model/ProductDetail.java
@@ -1,4 +1,4 @@
-package com.bidket.product.domain.model.ProductDetail;
+package com.bidket.product.domain.model;
 
 import java.util.UUID;
