+++ backend/product-service/src/main/java/com/bidket/product/infrastructure/persistence/entity/SizeType.java
@@ -1,4 +1,4 @@
-package com.bidket.product.domain.model;
+package com.bidket.product.infrastructure.persistence.entity;
 
 import com.bidket.common.infra.BaseEntity;
 import jakarta.persistence.Column;
@@ -17,30 +17,33 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @Table(
-        name = "p_size",
+        name = "p_size_type",
         indexes = {
-                @Index(name = "idx_size_size_type", columnList = "size_type_id"),
-                @Index(name = "idx_size_sort_id", columnList = "sort_id")
+                @Index(name = "idx_size_type_product_type", columnList = "product_type_id"),
+                @Index(name = "idx_size_type_code", columnList = "code")
         }
 )
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Size extends BaseEntity {
+public class SizeType extends BaseEntity {
 
     @Id
     @GeneratedValue
     private UUID id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "size_type_id", nullable = false)
-    private SizeType sizeType;
+    @JoinColumn(name = "product_type_id", nullable = false)
+    private ProductType productType;
 
     @Column(nullable = false, length = 50)
     private String code;
 
-    @Column(nullable = false, name = "display_label", length = 50)
-    private String displayLabel;
+    @Column(name = "region_code", length = 10)
+    private String regionCode;
 
-    @Column(nullable = false, name = "sort_id")
-    private Long sortId;
+    @Column(columnDefinition = "TEXT")
+    private String description;
+
+    @Column(nullable = false, name = "is_default")
+    private Boolean isDefault;
 }
