+++ backend/product-service/src/main/java/com/bidket/product/infrastructure/persistence/repository/ProductCategoryRepository.java
@@ -1,6 +1,6 @@
-package com.bidket.product.domain.repository;
+package com.bidket.product.infrastructure.persistence.repository;
 
-import com.bidket.product.domain.model.ProductCategory;
+import com.bidket.product.infrastructure.persistence.entity.ProductCategory;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
