+++ backend/product-service/src/main/java/com/bidket/product/infrastructure/persistence/repository/ProductSkuRepository.java
@@ -1,6 +1,6 @@
-package com.bidket.product.domain.repository;
+package com.bidket.product.infrastructure.persistence.repository;
 
-import com.bidket.product.domain.model.ProductSku;
+import com.bidket.product.infrastructure.persistence.entity.ProductSku;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
